requirements.txt
PyQt5==5.15.9
plyer==2.1.0
pystray==0.19.4
schedule==1.2.0
Pillow==9.3.0
